feat: generalized for handle mdoc cbor vps
